# api/openapi.yaml
openapi: 3.0.3
info:
  title: Big Project API
  version: 0.1.0
  description: >
    API for the COP4331 "Free Food" project. Covers authentication (registration
    with email verification, JWT-based login, password reset and change) and the
    food posts feature: creating posts, a confidence-ranked live feed,
    one-vote-per-user voting, and presigned image uploads. All endpoints are
    served under the `/api` prefix.

servers:
  - url: http://localhost:5001/api
    description: Local development

tags:
  - name: Health
    description: Service liveness
  - name: Auth
    description: Registration, login, verification, and password management
  - name: Posts
    description: Food posts feed, voting, and image uploads

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: getHealth
      security: []
      responses:
        '200':
          description: Service is up and the database is connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is up but the database is disconnected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      description: >
        Creates an unverified account and sends a verification email. The
        account cannot log in until the email is verified.
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created; verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Registered. Check your email to verify your account
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Email already in use

  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive a JWT
      description: Returns a 24h JWT. Rejects unverified accounts with 403.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid credentials
        '403':
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Email not verified

  /auth/verify:
    get:
      tags: [Auth]
      summary: Verify an email address
      description: >
        Consumes the token from the verification email. Returns an HTML page
        (this URL is opened directly from the email), not JSON.
      operationId: verifyEmail
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Raw verification token from the email link
      responses:
        '200':
          description: Email verified
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Missing, invalid, or expired token
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Unexpected error
          content:
            text/html:
              schema:
                type: string

  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend the verification email
      description: >
        Always returns 200 regardless of whether the email exists or is already
        verified, to avoid leaking which addresses are registered.
      operationId: resendVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Generic acknowledgement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: If that account exists and is unverified, a verification email has been sent.
        '400':
          $ref: '#/components/responses/ValidationError'

  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      description: >
        Always returns 200 regardless of whether the email exists, to avoid
        leaking which addresses are registered. Reset tokens expire after 1 hour.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Generic acknowledgement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: If that account exists, a password reset email has been sent.
        '400':
          $ref: '#/components/responses/ValidationError'

  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using a token
      description: >
        Consumes the token from the reset email and sets a new password. This
        invalidates all existing sessions; the user must log in again.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Password has been reset. You can now log in.
        '400':
          description: Missing fields, or invalid/expired reset token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid or expired reset token

  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change password while logged in
      description: >
        Requires the current password. On success, all other sessions are
        invalidated and a fresh JWT is returned so the calling client stays
        logged in. Clients must replace their stored token with the returned one.
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed; new token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Current password incorrect, or missing/invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Current password is incorrect
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: User not found

  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current user
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /posts:
    get:
      tags: [Posts]
      summary: List the live feed
      description: >
        Returns active posts (those not yet expired), most-alive-first. Each post
        includes a freshly computed `confidence` and `status`. Public; no auth.
      operationId: getFeed
      security: []
      responses:
        '200':
          description: The current feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
    post:
      tags: [Posts]
      summary: Create a food post
      description: >
        Creates a post seeded at a high starting confidence. If an image was
        uploaded via /posts/upload-url, pass the returned key as `imageKey`.
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePostResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /posts/upload-url:
    get:
      tags: [Posts]
      summary: Get a presigned image upload URL
      description: >
        Returns a short-lived (120s) presigned S3 PUT URL and the object key. The
        client PUTs the image bytes directly to `url`, then creates the post with
        `imageKey` set to `key`. Image bytes never pass through this API.
      operationId: getUploadUrl
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A presigned upload URL and its key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Image uploads are not configured (no S3 bucket set)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Image uploads are not configured

  /posts/{id}/vote:
    post:
      tags: [Posts]
      summary: Vote on a post
      description: >
        Records a "still here" or "gone" vote. Each user may vote once per post.
        The vote shifts the post's confidence and may expire it immediately.
      operationId: votePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The post's id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
      responses:
        '200':
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '400':
          description: Missing or invalid vote type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: type must be 'present' or 'gone'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Post not found
        '409':
          description: The user has already voted on this post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: You have already voted on this post

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by /auth/login or /auth/change-password. Send as
        `Authorization: Bearer <token>`. Tokens expire after 24h and are
        invalidated when the account's password changes.

  responses:
    ValidationError:
      description: A required field is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: email is required
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid or expired token

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    HealthResponse:
      type: object
      required: [status, uptime, db]
      properties:
        status:
          type: string
          enum: [ok, degraded]
          description: '"ok" when the database is connected, "degraded" otherwise.'
          example: ok
        uptime:
          type: number
          description: Process uptime in seconds
          example: 12.34
        db:
          type: string
          enum: [connected, disconnected]
          example: connected

    User:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          example: 665f1c2e9b1e8a0012a3b4c5
        email:
          type: string
          format: email
          example: user@example.com
        verified:
          type: boolean
          example: true

    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: hunter2

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: hunter2

    EmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: user@example.com

    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          description: Raw reset token from the email link
        password:
          type: string
          format: password
          description: The new password

    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT, valid for 24h
        user:
          $ref: '#/components/schemas/User'

    ChangePasswordResponse:
      type: object
      required: [token, message]
      properties:
        token:
          type: string
          description: Fresh JWT that replaces the caller's now-invalidated token
        message:
          type: string

    MeResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'

    PostStatus:
      type: string
      enum: [fresh, likely, fading, gone]
      description: Confidence band. Posts at "gone" are excluded from the feed.

    Tallies:
      type: object
      required: [present, gone]
      properties:
        present:
          type: integer
          example: 3
        gone:
          type: integer
          example: 1

    Post:
      type: object
      required: [_id, foodName, location, badges, author, status, tallies, expiresAt, createdAt, updatedAt]
      description: >
        A food post. Internal fields (e.g. the raw log-odds `E` and `lastUpdate`)
        may also appear in responses but are not part of the stable contract.
      properties:
        _id:
          type: string
          description: >
            MongoDB document id. Note: post objects expose `_id` (raw Mongoose
            shape), whereas user objects expose `id` — use the right one per resource.
          example: 665f1c2e9b1e8a0012a3b4c5
        foodName:
          type: string
          example: Leftover Domino's
        location:
          type: string
          example: HEC 101
        imageKey:
          type: string
          description: S3 object key; the client builds the display URL from this.
          example: posts/9661e87f-3b2c-4199-a797-4d91e7306955.jpg
        badges:
          type: array
          items:
            type: string
          example: [pizza, free]
        author:
          type: string
          description: Id of the user who posted it
          example: 665f1c2e9b1e8a0012a3b4c5
        status:
          $ref: '#/components/schemas/PostStatus'
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: >
            Live probability the food is still there (0–1). Computed on read, so
            it is present on feed responses.
          example: 0.88
        tallies:
          $ref: '#/components/schemas/Tallies'
        expiresAt:
          type: string
          format: date-time
          description: When the post will decay past the "gone" threshold.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreatePostRequest:
      type: object
      required: [foodName, location]
      properties:
        foodName:
          type: string
          example: Leftover Domino's
        location:
          type: string
          example: HEC 101
        badges:
          type: array
          items:
            type: string
          example: [pizza, free]
        imageKey:
          type: string
          description: Key from /posts/upload-url, if an image was uploaded.

    CreatePostResponse:
      type: object
      required: [post]
      properties:
        post:
          $ref: '#/components/schemas/Post'

    FeedResponse:
      type: object
      required: [posts]
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/Post'

    VoteRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [present, gone]
          description: '"present" = still here, "gone" = not there anymore'

    VoteResponse:
      type: object
      required: [confidence, status, tallies]
      properties:
        confidence:
          type: number
          format: float
          example: 0.27
        status:
          $ref: '#/components/schemas/PostStatus'
        tallies:
          $ref: '#/components/schemas/Tallies'

    UploadUrlResponse:
      type: object
      required: [url, key]
      properties:
        url:
          type: string
          format: uri
          description: Presigned S3 PUT URL, valid for 120 seconds.
        key:
          type: string
          example: posts/9661e87f-3b2c-4199-a797-4d91e7306955.jpg